0507 1.refine dns switch. 2.refine tips. 3.allow disable mu in some node. 4.refine traffic log. 5.allow set none method when some protocol is enable.

+++ app/Models/User.php
@@ -26,9 +26,9 @@ class User extends Model
     public $isAdmin;
 
     protected $casts = [
-        "t" => 'int',
-        "u" => 'int',
-        "d" => 'int',
+        "t" => 'float',
+        "u" => 'float',
+        "d" => 'float',
         "port" => 'int',
         "transfer_enable" => 'float',
         "enable" => 'int',
